handle fetchStandardFontData error
When parsing a document that attempts to reach a new font file from the standard font path, it throws a warning:
```
Warning: fetchStandardFontData: failed to fetch file "FoxitSans.pfb" with "UnknownErrorException:  The standard font "baseUrl" parameter must be specified, ensure that the "standardFontDataUrl" API parameter is provided.".
```

This automatically passes in the standardFontDataUrl path with the default path of the pdfjs-dist package. A thought had occurred to pass this in via argument but it's already very verbose and this path should be standard across most machines using pdf2md.